Update for README.md
This is a small linux application for quickly editing the control file for a .deb package. Only software required is a .deb package and a text editor. To change what text editor you use simply open the dcfe file and change it. The default is leafpad. to install simple place thee dcfe file in /usr/bin of your linux system.